+++ examples/resources/appleads_ad_group/resource.tf
@@ -8,5 +8,15 @@ resource "appleads_ad_group" "main" {
   status                    = "PAUSED"
   default_bid_amount        = "1.25"
   default_bid_currency      = "USD"
+  pricing_model             = "CPC"
   automated_keywords_opt_in = true
+}
+
+resource "appleads_ad_group" "today_tab" {
+  campaign_id          = appleads_campaign.display.id
+  name                 = "Today Tab"
+  status               = "PAUSED"
+  default_bid_amount   = "5.00"
+  default_bid_currency = "USD"
+  pricing_model        = "CPM"
 }